Update push feature
