+++ app/javascript/dashboard/i18n/locale/en/settings.json
@@ -10,11 +10,11 @@
       "REMOVE_IMAGE": "Remove",
       "UPLOAD_IMAGE": "Upload image",
       "UPDATE_IMAGE": "Update image",
-      "PROFILE_SECTION" : {
+      "PROFILE_SECTION": {
         "TITLE": "Profile",
         "NOTE": "Your email address is your identity and is used to log in."
       },
-      "PASSWORD_SECTION" : {
+      "PASSWORD_SECTION": {
         "TITLE": "Password",
         "NOTE": "Updating your password would reset your logins in multiple devices."
       },
@@ -22,7 +22,7 @@
         "TITLE": "Access Token",
         "NOTE": "This token can be used if you are building an API based integration"
       },
-      "EMAIL_NOTIFICATIONS_SECTION" : {
+      "EMAIL_NOTIFICATIONS_SECTION": {
         "TITLE": "Email Notifications",
         "NOTE": "Update your email notification preferences here",
         "CONVERSATION_ASSIGNMENT": "Send email notifications when a conversation is assigned to me",
@@ -32,13 +32,15 @@
         "UPDATE_SUCCESS": "Your notification preferences are updated successfully",
         "UPDATE_ERROR": "There is an error while updating the preferences, please try again"
       },
-      "PUSH_NOTIFICATIONS_SECTION" : {
+      "PUSH_NOTIFICATIONS_SECTION": {
         "TITLE": "Push Notifications",
         "NOTE": "Update your push notification preferences here",
         "CONVERSATION_ASSIGNMENT": "Send push notifications when a conversation is assigned to me",
-        "CONVERSATION_CREATION": "Send push notifications when a new conversation is created"
+        "CONVERSATION_CREATION": "Send push notifications when a new conversation is created",
+        "HAS_ENABLED_PUSH": "You have enabled push for this browser.",
+        "REQUEST_PUSH": "Enable push notifications"
       },
-      "PROFILE_IMAGE":{
+      "PROFILE_IMAGE": {
         "LABEL": "Profile Image"
       },
       "NAME": {
