+++ app/javascript/dashboard/routes/dashboard/settings/profile/Index.vue
@@ -82,7 +82,7 @@
           </label>
         </div>
       </div>
-      <email-notifications />
+      <notification-settings />
       <div class="profile--settings--row row">
         <div class="columns small-3 ">
           <h4 class="block-title">
@@ -101,26 +101,6 @@
       >
       </woot-submit-button>
     </form>
-    <div class="profile--settings--row row">
-      <div class="columns small-3 ">
-        <h4 class="block-title">
-          Push Notifications
-        </h4>
-        <p>Enable push notifications on this browser</p>
-      </div>
-      <div class="columns small-9 medium-5">
-        <p v-if="hasEnabledPushPermissions">
-          You have enabled push for this browser.
-        </p>
-        <div v-else>
-          <woot-submit-button
-            class="button nice"
-            button-text="Request push permission"
-            @click="requestPushPermissions"
-          />
-        </div>
-      </div>
-    </div>
   </div>
 </template>
 
@@ -131,15 +111,11 @@ import Thumbnail from 'dashboard/components/widgets/Thumbnail.vue';
 import { required, minLength, email } from 'vuelidate/lib/validators';
 import { mapGetters } from 'vuex';
 import { clearCookiesOnLogout } from '../../../../store/utils/api';
-import EmailNotifications from './EmailNotifications';
-import {
-  hasPushPermissions,
-  requestPushPermissions,
-} from '../../../../helper/pushHelper';
+import NotificationSettings from './NotificationSettings';
 
 export default {
   components: {
-    EmailNotifications,
+    NotificationSettings,
     Thumbnail,
   },
   data() {
@@ -179,9 +155,6 @@ export default {
       currentUser: 'getCurrentUser',
       currentUserId: 'getCurrentUserID',
     }),
-    hasEnabledPushPermissions() {
-      return hasPushPermissions();
-    },
   },
   watch: {
     currentUserId(newCurrentUserId, prevCurrentUserId) {
@@ -196,7 +169,6 @@ export default {
     }
   },
   methods: {
-    requestPushPermissions,
     initializeUser() {
       this.name = this.currentUser.name;
       this.email = this.currentUser.email;
